% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_multi.R
\name{create_betas}
\alias{create_betas}
\title{Generate a beta vector}
\usage{
create_betas(maf, genetic_var, ncausal)
}
\arguments{
\item{maf}{allele frequencies of the variants of interest}

\item{genetic_var}{the range of variation due to genetic effect}

\item{ncausal}{the range of the number of causal variants}
}
\value{
a vector of simulated beta vector
}
\description{
This functions simulate a beta vector with the variation of genetic effect
being controlled in a pre-specified range.
It takes maf vector of length P, range of desired variation due to genetic effect, and
the range of the number of causal variants.
The variation is calculated by \deqn{2 \sum_p \beta_p^2 f_p (1 - f_p)} where \eqn{f_p} is the allele frequency of variant p and \eqn{\beta_p} is the log aFC of variant p.
}
\examples{
create_betas(
  maf = runif(100),
  genetic_var = c(0.015, 0.075),
  ncausal = c(1, 3)
)

}
