% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_simulation.R
\name{create_gene}
\alias{create_gene}
\title{Generate a gene instance for simulation}
\usage{
create_gene(
  L_gene = 10000,
  library_dist = list(type = "negbinom", prob = 1.6e-07, size = 15),
  theta_dist = list(type = "beta", mu = 1e-05, sd = 5e-06),
  f = 0.001,
  maf = c(0.05, 0.3),
  p = rep(1/10000, 10000)
)
}
\arguments{
\item{L_gene}{length of gene body}

\item{library_dist}{distribution of library size, a list of 'type' (can be either 'poisson' or 'negbinom') and corresponding parameters: 1. 'negbinom': 'prob' and 'size'; 2. 'poisson': 'lambda'}

\item{theta_dist}{distribution of baseline abundance theta_{0,i}, a list of 'type' (can be either 'beta' or 'lognormal') and corresponding parameters: 1. 'beta': 'alpha' and 'beta'; 2. 'lognormal': 'k' (mu) and 'sigma'}

\item{f}{frequency of polymorphic site along the gene body}

\item{maf}{allele frequency range of polymorphic site (MAF ~ Unif(maf_low, maf_high))}

\item{p}{probability of having read at each position along the gene body}
}
\value{
an gene instance
}
\description{
This function generates a gene region to work with.
For a gene region, the following parameters are required
  1. L_gene: length of gene region;
  2. library_dist: distribution of library size (e.g. NegBionmial and Poisson);
  3. theta_dist: distribution of theta_{0,i} (e.g. Beta);
  4. f: frequency of polymorphic site along gene body;
  5. maf: MAF (a range) of polymorphic site;
  6. p: probability of having reads by position.
}
\examples{
create_gene()

}
