% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_simulation.R
\name{create_genotype}
\alias{create_genotype}
\title{Generate genotype (for single SNP model only)}
\usage{
create_genotype(maf, nsample, nreplicate)
}
\arguments{
\item{maf}{MAF range, c(maf_low, maf_high)}

\item{nsample}{number of individual to simulate}

\item{nreplicate}{number of replicates to simulation}
}
\value{
a list of length nreplicate where each element contains a data.frame of simulation genotype (for each haplotype) along with allele frequency.
}
\description{
Given MAF range and sample size, generate dosage genotype (0, 1, 2) for one site.
}
\examples{
create_genotype(
  maf = c(0.05, 0.3),
  nsample = 300,
  nreplicate = 1000
)
}
