% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_multi_util.R
\name{do_elastic_net}
\alias{do_elastic_net}
\title{Adapted from https://github.com/hakyimlab/PredictDB_Pipeline_GTEx_v7/blob/master/model_training/scripts/gtex_v7_elasticnet.R}
\usage{
do_elastic_net(cis_gt, expr_adj, n_folds, cv_fold_ids, alpha, ...)
}
\arguments{
\item{cis_gt}{x (genotype matrix, sample x variant)}

\item{expr_adj}{y (expression level vector)}

\item{n_folds}{cross-validation fold}

\item{cv_fold_ids}{K partitions with N-fold (sample x K, where each column is partition of samples into N parts represented by the index of each part)}

\item{alpha}{alpha parameter in glmnet}

\item{...}{additional arguments passed to cv.glmnet}
}
\description{
Fit y ~ x + 1 with elastic net using nested cross-validation.
By "nested", we mean determine MSE using multiple partitions (K)
instead of only one in vanilla cross-validation.
The procedure is:
1. determine lambda sequence and fit y ~ x at different lambda values;
2. take in K copies of the partition of the input (one of the input, cv_fold_ids).
For instance, 5-fold partition if you'd like 5-fold CV;
3. compute CV MSE at each lambda value for each fold in each partition;
4. compute CV MSE for each lambda as the mean of CV MSE of all partitions;
5. select the best model with the smallest CV MSE at that lambda.
}
