% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_multi_util.R
\name{fit_glmnet_with_cv}
\alias{fit_glmnet_with_cv}
\title{Wrapper for fitting elastic net model with nested cross-validation}
\usage{
fit_glmnet_with_cv(
  x,
  y,
  nfold = 5,
  lambda_seq = NULL,
  alpha = 0.5,
  n_times = 3,
  ...
)
}
\arguments{
\item{x}{x (N by P, usually P >> N)}

\item{y}{y (N by 1)}

\item{nfold}{fold in cross-valiation to pick hyperparameter lambda (default = 5)}

\item{lambda_seq}{specify the sequence of hyperparameter lambda to search.
If don't have such lambda_seq, set it to NULL,
and the function will generate one internally (default = NULL)}

\item{alpha}{alpha parameter in glmnet (defualt = 0.5)}

\item{n_times}{how many rounds of partitions you'd like to perform to estimate
CV MSE for each lambda (it is specific for nested CV procedure; default = 3)}

\item{...}{additional arguments passed to cv.glmnet}
}
\description{
Build additive linear model to predict y from x using elastic net.
It fits the use case when the number of predictors is much larger than
the number of observations.
}
\seealso{
\code{\link{do_elastic_net}} for details on nested CV procedure.
}
