% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_util.R
\name{impute_geno}
\alias{impute_geno}
\title{Filling in missing genotype}
\usage{
impute_geno(x)
}
\arguments{
\item{x}{genotype (dimension N x P)}
}
\value{
genotype matrix with missing values being imputed
}
\description{
Filling in the missing genotype as the average.
}
\examples{
impute_geno(
  x = matrix(sample(c(0, 0.5, 1), 200, replace = TRUE), ncol = 2)
)

}
