% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_matrix_ls.R
\name{matrix_ls_asc}
\alias{matrix_ls_asc}
\title{ascQTL for nominal pass in matrix form}
\usage{
matrix_ls_asc(asc1, asc2, x, asc_cutoff = 5, weight_cap = 100, asc_cap = 5000)
}
\arguments{
\item{asc1}{allele-specific read count for haplotype 1 (dimension = N x 1)}

\item{asc2}{allele-specific read count for haplotype 2 (dimension = N x 1)}

\item{x}{'genotype' used in ascQTL (which is defined as x1 - x2) (dimension = N x P)}

\item{asc_cutoff}{allele-specific read count cutoff to exclude observations with asc1 or asc2 lower than asc_cutoff}

\item{weight_cap}{the maximum weight difference (in fold) is min(weight_cap, floor(sample_size / 10)). The ones exceeding the cutoff is capped.}

\item{asc_cap}{exclude observations with asc1 or asc2 higher than asc_cap}
}
\value{
a list of summary statistics
        beta_hat: estimated log aFC (dimension = 1 x P)
        beta_se: standard deviation of log aFC (dimension = 1 x P)
        sample_size: number of observations used in regression
}
\description{
Perform ascQTL in matrix form which can be used for obtaining nominal p-value
}
\examples{
matrix_ls_asc(
  asc1 = rpois(100, 100),
  asc2 = rpois(100, 100),
  x = matrix(sample(c(0, 0.5, 1), 200, replace = TRUE), ncol = 2),
  asc_cutoff = 20,
  weight_cap = 10,
  asc_cap = 1000
)

}
