% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_matrix_ls.R
\name{matrix_ls_trc}
\alias{matrix_ls_trc}
\title{trcQTL for nominal pass in matrix form}
\usage{
matrix_ls_trc(trc, lib_size, x, cov, trc_cutoff = 20)
}
\arguments{
\item{trc}{total read count (dimension = N x 1)}

\item{lib_size}{library size (dimension = N x 1)}

\item{x}{'genotype' used in trcQTL (which is defined as (x1 + x2) / 2) (dimension = N x P)}

\item{cov}{effect of covariates on log(trc/2) (dimension = N x 1)}

\item{trc_cutoff}{total read count cutoff to exclude observations with trc lower than the cutoff}
}
\value{
a list of summary statistics
        beta_hat: estimated log aFC (dimension = 1 x P)
        beta_se: standard deviation of log aFC (dimension = 1 x P)
        sample_size: number of observations used in regression
}
\description{
Perform trcQTL in matrix form which can be used for obtaining nominal p-value
}
\examples{
matrix_ls_trc(
  trc = rpois(100, 100),
  lib_size = rpois(100, 10000),
  x = matrix(sample(c(0, 0.5, 1), 200, replace = TRUE), ncol = 2),
  cov = rnorm(100),
  trc_cutoff = 20
)

}
