% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_matrix_ls.R
\name{matrixQTL_two_dim}
\alias{matrixQTL_two_dim}
\title{Solve Y = X1 b1 + X2 b2 + e in matrix form}
\usage{
matrixQTL_two_dim(Y, X1, X2, n)
}
\arguments{
\item{Y}{response to regress against (dimension = N x K)}

\item{X1}{P predictors (as the first predictor) to perform regression separately (dimension = N x P)}

\item{X2}{P predictors (as the second predictor) to perform regression separately (dimension = N x P)}

\item{n}{sample size (dimension = P x K)}
}
\value{
a list of summary statistics
        beta1_hat: estimated b1, b1_hat (dimension = K x P)
        beta1_se: standard deviation of b1_hat (dimension = K x P)
        beta2_hat: estimated b2, b2_hat (dimension = K x P)
        beta2_se: standard deviation of b2_hat (dimension = K x P)
}
\description{
For each column i in X, solve Y_k = X_1i b_1i + X_2i b_2i + e as least squares problem and output estimated effect size and standard deviation
}
\examples{
matrixQTL_two_dim(
  Y = matrix(rnorm(100), ncol = 3),
  X1 = matrix(sample(c(0, 0.5, 1), 200, replace = TRUE), ncol = 2),
  X2 = matrix(sample(c(0, 0.5, 1), 200, replace = TRUE), ncol = 2),
  n = matrix(200, ncol = 2, nrow = 3)
)

}
