% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixqtl.R
\name{mixqtl}
\alias{mixqtl}
\title{Combining total and allele-specific reads for QTL mapping}
\usage{
mixqtl(
  geno1,
  geno2,
  y1,
  y2,
  ytotal,
  lib_size,
  cov_offset = NULL,
  trc_cutoff = 20,
  asc_cutoff = 5,
  weight_cap = 100,
  asc_cap = 5000
)
}
\arguments{
\item{geno1}{genotype of haplotype 1 (dimension = N x P)}

\item{geno2}{genotype of haplotype 2 (dimension = N x P)}

\item{y1}{allele-specific read count for haplotype 1 (dimension = N x 1)}

\item{y2}{allele-specific read count for haplotype 2 (dimension = N x 1)}

\item{ytotal}{total read count (dimension = N x 1)}

\item{lib_size}{library size (dimension = N x 1)}

\item{cov_offset}{predicted effect of covariates on total read count response term,
log(ytotal / lib_size / 2) (dimension = N x 1)}

\item{trc_cutoff}{total read count cutoff to exclude observations with ytotal lower than the cutoff}

\item{asc_cutoff}{allele-specific read count cutoff to exclude observations with y1 or y2 lower than asc_cutoff}

\item{weight_cap}{the maximum weight difference (in fold) is min(weight_cap, floor(sample_size / 10)). The ones exceeding the cutoff is capped.}

\item{asc_cap}{exclude observations with y1 or y2 higher than asc_cap}
}
\value{
QTL mapping result
}
\description{
Given total and allele-specific read counts along with library size,
MixQTL procedure (by meta-analysis) is implemented.
Specifically, first, ascQTL and trcQTL estimates are obtained.
Then, second, meta-analyze ascQTL and trcQTL estimates to obtain mixQTL estimate.
The function outputs the estimates of outcome versus each of the input variants
}
\examples{
mixqtl(
  geno1 = matrix(sample(c(0, 0.5, 1), 200, replace = TRUE), ncol = 2),
  geno2 = matrix(sample(c(0, 0.5, 1), 200, replace = TRUE), ncol = 2),
  y1 = rpois(100, 100),
  y2 = rpois(100, 100),
  ytotal = rpois(100, 1000),
  lib_size = rpois(100, 10000),
  cov_offset = runif(100),
  trc_cutoff = 100,
  asc_cutoff = 50,
  weight_cap = 100,
  asc_cap = 1000
)

}
