% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_covariate.R
\name{regress_against_covariate}
\alias{regress_against_covariate}
\title{Regress against covariates for trcQTL}
\usage{
regress_against_covariate(trc, lib_size, covariates)
}
\arguments{
\item{trc}{total read count (dimension = N x 1)}

\item{lib_size}{library size (dimension = N x 1)}

\item{covariates}{covariates (dimension = C x N) with the first column indicating the covariate name}
}
\value{
predicted response based on the selected covariates (if no covariates are selected, fill in zeros)
}
\description{
Given total read count (trc, N x 1) and library size (lib_size, N x 1) along with covarites (N x C),
regress covariates against log(trc / lib_size / 2) by two-step approach
1. Regress all covariates against response and select significant ones (alpha = 0.05)
2. Regress the selected covariates against response and return the predicted response
}
\examples{
regress_against_covariate(
  trc = rpois(100, 100),
  lib_size = rpois(100, 10000),
  covariates = cbind(
    data.frame(name = paste0('cov', 1:10)),
    as.data.frame(matrix(runif(1000), nrow = 10))
  )
)

}
