% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_matrix_ls.R
\name{meta_analyze}
\alias{meta_analyze}
\title{Solve Y = X1 b1 + X2 b2 + e in matrix form}
\usage{
meta_analyze(trc, asc)
}
\arguments{
\item{trc}{list of effect size estimates and standard deviation (e.g. output of matrix_ls_trc)}

\item{asc}{list of effect size estimates and standard deviation (e.g. output of matrix_ls_asc)}
}
\value{
a list of summary statistics and p-values obtained from trc, asc, and meta-analysis of the two
        the object within list is named by the method (trc, asc, and meta) and each of which includes
        pval (p-value), stat (test statistic), stat_type (z-value or t-value),
        bhat (effect size estimate), se (standard deviation of effect size),
        method (only for 'meta' object showing if the result is from 'trc', 'asc', or 'meta')
}
\description{
For each column i in X, solve Y = X_1i b_1i + X_2i b_2i + e as least squares problem and output estimated effect size and standard deviation
}
\examples{
meta_analyze(
  trc = list(
    beta_hat = matrix(rnorm(100), ncol = 20),
    beta_se = abs(matrix(rnorm(100), ncol = 20)),
    sample_size = 100
  ),
  asc = list(
    beta_hat = matrix(rnorm(100), ncol = 20),
    beta_se = abs(matrix(rnorm(100), ncol = 20)),
    sample_size = 100
  )
)

}
